COMMON.ENH: Add EpicsWait thread to wait for PVs to update.
