+++ pyqt-apps/siriushla/common/epics/task/__init__.py
@@ -4,3 +4,4 @@ from .getter import EpicsGetter
 from .setter import EpicsSetter
 from .checker import EpicsChecker
 from .connector import EpicsConnector
+from .wait import EpicsWait
